// training_frontend/src/api/models/Workout.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Training service
 * Backend for training app
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { WorkoutExercise } from './WorkoutExercise';
import {
    WorkoutExerciseFromJSON,
    WorkoutExerciseFromJSONTyped,
    WorkoutExerciseToJSON,
    WorkoutExerciseToJSONTyped,
} from './WorkoutExercise';
import type { MuscleGroup } from './MuscleGroup';
import {
    MuscleGroupFromJSON,
    MuscleGroupFromJSONTyped,
    MuscleGroupToJSON,
    MuscleGroupToJSONTyped,
} from './MuscleGroup';

/**
 * 
 * @export
 * @interface Workout
 */
export interface Workout {
    /**
     * 
     * @type {string}
     * @memberof Workout
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Workout
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof Workout
     */
    editedAt: number;
    /**
     * 
     * @type {Array<MuscleGroup>}
     * @memberof Workout
     */
    muscleGroups: Array<MuscleGroup>;
    /**
     * 
     * @type {Array<WorkoutExercise>}
     * @memberof Workout
     */
    exercises: Array<WorkoutExercise>;
}

/**
 * Check if a given object implements the Workout interface.
 */
export function instanceOfWorkout(value: object): value is Workout {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('editedAt' in value) || value['editedAt'] === undefined) return false;
    if (!('muscleGroups' in value) || value['muscleGroups'] === undefined) return false;
    if (!('exercises' in value) || value['exercises'] === undefined) return false;
    return true;
}

export function WorkoutFromJSON(json: any): Workout {
    return WorkoutFromJSONTyped(json, false);
}

export function WorkoutFromJSONTyped(json: any, ignoreDiscriminator: boolean): Workout {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'editedAt': json['editedAt'],
        'muscleGroups': ((json['muscleGroups'] as Array<any>).map(MuscleGroupFromJSON)),
        'exercises': ((json['exercises'] as Array<any>).map(WorkoutExerciseFromJSON)),
    };
}

export function WorkoutToJSON(json: any): Workout {
    return WorkoutToJSONTyped(json, false);
}

export function WorkoutToJSONTyped(value?: Workout | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'editedAt': value['editedAt'],
        'muscleGroups': ((value['muscleGroups'] as Array<any>).map(MuscleGroupToJSON)),
        'exercises': ((value['exercises'] as Array<any>).map(WorkoutExerciseToJSON)),
    };
}